Big Changes!
Reorganized the whole extension!

Joined the main categories into one, diferenced by colors and labels.

Currently editing the addons. Thinking if they should go together in one file or separated.

I did all of this to fix the flood issue. With it I had to add _Extra3D_ as the API, so the addons file could speak to the main.

If you use the addons file now, you will get flood.